Bugfixes
Moved _printBoard method from human to player so that the AI class can print the board.
FIxes to the medium AI as regardless if it was on the WHITE or BLACK team it would always try to maximize the score. It now either will attempt to maximize or minimize the score based on self.assignedPiece 's value.
Several small bugfixes in AI.py