Loading saving system fix + Exception handling
Game class-
Adding a delete function to the checkIfFileExists method to delete the temorary file created when checking if the file exists.
Small fixes around the code, -dealWithSpecialCondition had a couple of syntax errors.
Adding a try: except Exception error to the main menu to handle any unhandled exceptions in the program and report them to the user.
AI-
Adding a try: except Exception line to the getMove method that returns false when an error occurs and reports an error to the user.