Addition of rules to main menu
Adds a small bit of text in the game class in a private method called _tutorial that prints out rules for the game explaining to the user how it works.
Small tweaks to the UI, including the addition of the globalfunctions.typing function that types out the next character by character in the console instead of all at once.