Fixes to AI.py
Fixes to the alpha-beta pruning part of the minimax algorithm. Instead of the maxHeuristic and minHeuristic move being used, the statement is replaced with beta >= alpha.
Moving the break statement for the alpha beta pruning to the front of the for loop so that it will not search through the tree at all and will terminate instead of searching one node if the condition is met.
Addition of an extra condition to the heuristic calculation- if the game is a draw, return heuristic 0.